Логика работы с фильмом реализована, необходимо тестирование

+++ KINOv2/KINOv2/Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -204,7 +204,7 @@ namespace KINOv2.Data.Migrations
 
                     b.Property<int?>("FilmLINK");
 
-                    b.Property<int>("Value");
+                    b.Property<double>("Value");
 
                     b.HasKey("LINK");
